requirements.txt
awsiotsdk
pyserial
pymodbus
nidaqmx